// src/peo-upload/dto/upload-pegawai.dto.ts
export class UploadPegawaiDto {
  KODE_SAP: string;
  NAMA: string;
  NIPP: string;
  KD_CABANG_SAP: string;
  NAMA_CABANG: string;
  KD_JABATAN: string;
  NAMA_JABATAN: string;
  KELAS_JABATAN: string;
  KD_DIR: string;
  NAMA_DIR: string;
  KD_SUB: string;
  NAMA_SUB: string;
  EMAIL: string;
  HP: string;
  ACCOUNT_AD: string;
  KD_WIL_ARSIP: string;
  KD_DIV_ARSIP: string;
  COOKIE: string;
  PERSG: string;
  KD_PEL: string;
  SUB_AREA: string;
  NAMA_SUB_AREA: string;
  JENIS: string;
  NOTIF: string;
  JENIS_SK: string;
  NAMA_JABATAN_SK: string;
  TRFS0: string;
  TMT_PERIODIK: string;
  TUNJANGAN: string;
  TALHIR: string;
  PATHIR: string;
  STEXT: string;
  TTL: string;
  KODE_NOMOR: string;
  AKSES_SURAT: string;
  NAMA_DIREKTORAT: string;
  KELOMPOK: string;
  NIPP_BARU: string;
  TTD: string;
  CABANG_NAME: string;
  GRUP: string;
  PENGOLAH: string;
  BISA_HAPUS: string;
  COMPANY_CODE: string;
  INSTANSI: string;
  LAST_UPDATED_DATE: string;
  COST_CENTER: string;
  BANK_KEY: string;
  BANK_NAME: string;
  BANK_ACCOUNT: string;
  WERKS_NEW: string;
  PBTXT_NEW: string;
  BTRTL_NEW: string;
  BTRTX_NEW: string;
  CREATED_DATE: string;
  TRAVELCOSTCENTER: string;
  NAMA_SUB_TRAVEL: string;
  ID: string;
  COSTCENTER_STO: string;
  COSTCENTERTEXT_STO: string;
  CHIEFPOSITION: string;
  PERSA_STO: string;
  SUBPERSA_STO: string;
  PERSATEXT_STO: string;
  SUBPERSATEXT_STO: string;
  PAYSCALETYPE: string;
  PAYSCALETYPETEXT: string;
  PEGAWAI: string;
  KODE_DIREKTORAT: string;
}